refactor: Change Refresh token service package
